[AIP-21] Move contrib.operators.qubole_operator providers.qubole.operators.qubole
