+++ airflow/contrib/example_dags/example_qubole_operator.py
@@ -32,9 +32,9 @@ import filecmp
 import random
 
 from airflow import DAG
-from airflow.contrib.operators.qubole_operator import QuboleOperator
 from airflow.operators.dummy_operator import DummyOperator
 from airflow.operators.python import BranchPythonOperator, PythonOperator
+from airflow.providers.qubole.operators.qubole import QuboleOperator
 from airflow.utils.dates import days_ago
 
 default_args = {
