+++ docs/operators-and-hooks-ref.rst
@@ -943,7 +943,7 @@ These integrations allow you to perform various operations within various servic
      -
      - :mod:`airflow.providers.qubole.hooks.qubole`,
        :mod:`airflow.providers.qubole.hooks.qubole_check`
-     - :mod:`airflow.contrib.operators.qubole_operator`,
+     - :mod:`airflow.providers.qubole.operators.qubole`,
        :mod:`airflow.providers.qubole.operators.qubole_check`
      - :mod:`airflow.contrib.sensors.qubole_sensor`
